memory/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
chromadb==0.4.18
sentence-transformers==2.2.2
torch==2.1.0
pillow==10.1.0
requests==2.31.0